project_root/requirements.txt
google-play-scraper
pyyaml
pandas
scikit-learn
Sastrawi
numpy==1.26.4
gensim==4.3.3
nltk
torch
huggingface
transformers